organized components a bit and added button to return to dashboard after adding food items to an event
