--- src/components/Users/UserCard.js
@@ -1,9 +1,0 @@
-import React from 'React';
-
-function UserCard () {
-    
-    return (
-        <h1>card</h1>
-    );
-}
-export default UserCard
