--- src/components/Users/UserInfo.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-function UserInfo () {
-    
-    return (
-        <h1>User Info</h1>
-    );
-}
-export default UserInfo
